polars_sql
Module keywords

Keywords that are supported by the Polars SQL interface.

This module defines:

all recognised Polars SQL keywords all_keywords

all recognised Polars SQL functions all_functions

Functions

all_functions

Get a list of all function names that are supported by Polars SQL

all_keywords

Get all keywords that are supported by Polars SQL